- solved warning
git-svn-id: http://galileo.dmi.unict.it/svn/weborf/trunk@798 27c2e2b3-ca6d-4234-b4fe-3aff1b8b8c5a

+++ configuration.c
@@ -348,7 +348,7 @@ const char * configuration_get_cgi(connection_t * connection_prop) {
 
     if (weborf_conf.exec_script) { //Scripts enabled
         size_t q_;
-        int f_len;
+        size_t f_len;
         for (q_=0; q_<weborf_conf.cgi_paths.len; q_+=2) { //Check if it is a CGI script
 
             f_len=weborf_conf.cgi_paths.data_l[q_];
